Make commitStyles use computed values instead of resolved values
Add new CSSAnimationUtils class with a static method
KeyframeValueFromComputedStyle() which returns the true computed value
for all properties. This mostly calls CSSValueForComputedStyle() without
a LayoutObject, but the cases where this does not return the computed
value (line-height and transform) are handles separately, similarly
to the implementation of ComputedStyleMap.

This method does take a LayoutObject, but it is only used to resolve
matrix interpolation of box-size-dependent transform values, which
produces a deferred compilation which cannot be serialized directly in
the current spec.

Make Animation.commitStyles() use this new function instead of
committing the resolved value using CSSValueFromComputedStyle().

Add additional WPT tests for commitStyles() adding tests for
special-case properties and splitting the logical properties case.

Bug: 1056197, 1057307
Change-Id: Ic823f96f257d4c6b010dfee758d4ac0b5cd4148c